+++ integrations/aws/utils/config.py
@@ -1,19 +1,24 @@
 import typing
-from integration import AWSPortAppConfig, AWSResourceConfig
+from aws.override import AWSPortAppConfig, AWSResourceConfig
 from port_ocean.context.event import event
 from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 
 
-def get_matching_kinds_from_config(kind: str) -> list[ResourceConfig]:
-    return list(
-        set(
-            filter(
-                lambda resource_config: resource_config.kind == kind
-                or (
-                    isinstance(resource_config, AWSResourceConfig)
-                    and kind in resource_config.selector.resource_kinds
-                ),
-                typing.cast(AWSPortAppConfig, event.port_app_config).resources,
-            )
-        )
-    )
+def get_matching_kinds_from_config(
+    kind: str,
+) -> list[ResourceConfig | AWSResourceConfig]:
+    result: list[ResourceConfig | AWSResourceConfig] = []
+    kinds_cache: list[str] = []
+    resources = typing.cast(AWSPortAppConfig, event.port_app_config).resources
+
+    for resource in resources:
+        if resource.kind in kinds_cache:
+            continue
+        if (
+            isinstance(resource, AWSResourceConfig)
+            and kind in resource.selector.resource_kinds
+        ) or kind == resource.kind:
+            result.append(resource)
+            kinds_cache.append(resource.kind)
+
+    return result
